Find Switches (#108)
