+++ generate-osm/db-utils/map-signals.go
@@ -3,8 +3,6 @@ package dbUtils
 import (
 	"encoding/xml"
 	"fmt"
-	"math"
-	"sort"
 	"strconv"
 	"strings"
 	OSMUtil "transform-osm/osm-utils"
@@ -91,9 +89,7 @@ func processHauptsignal(
 			}
 		}
 
-		if len(matchingSignalNodes) > 1 {
-			(*conflictingSignalNames)[signal.Name.Value] = true
-		} else if len(matchingSignalNodes) == 1 {
+		if len(matchingSignalNodes) == 1 {
 			conflictFreeSignal, err := insertNewHauptsignal(
 				optionalNewId,
 				matchingSignalNodes[0],
@@ -112,8 +108,8 @@ func processHauptsignal(
 				*foundAnchorCount++
 				return nil
 			}
-			(*conflictingSignalNames)[signal.Name.Value] = true
 		}
+		(*conflictingSignalNames)[signal.Name.Value] = true
 		*notFoundSignals = append(*notFoundSignals, signal)
 	}
 	return nil
@@ -164,46 +160,25 @@ func insertNewHauptsignal(
 			}
 		}
 	}
+
 	newSignalNode := createNewHauptsignal(
 		newId,
 		signalNode,
 		signal,
 		isFalling,
 	)
-	OSMUtil.InsertNewNodeWithReferenceNode(osm, &newSignalNode, signalNode)
+	OSMUtil.InsertNewNodeWithReferenceNode(
+		osm,
+		&newSignalNode,
+		signalNode,
+	)
+
 	if len(anchors[signalKilometrage]) == 0 {
 		anchors[signalKilometrage] = []*OSMUtil.Node{&newSignalNode}
 	} else {
 		anchors[signalKilometrage] = append(anchors[signalKilometrage], &newSignalNode)
 	}
 	return true, nil
-}
-
-// createNewHauptsignal creates a new OSM-Node with the following tags:
-// 'type:element', 'subtype:ms', 'id:(Signal name)' and 'direction:...' where ... depends on 'isFalling'.
-func createNewHauptsignal(
-	id *int,
-	node *OSMUtil.Node,
-	signal *Signal,
-	isFalling bool,
-) OSMUtil.Node {
-	directionString := "falling"
-	if !isFalling {
-		directionString = "rising"
-	}
-	*id++
-
-	return OSMUtil.Node{
-		Id:  strconv.Itoa(*id),
-		Lat: node.Lat,
-		Lon: node.Lon,
-		Tag: []*OSMUtil.Tag{
-			{XMLName: XML_TAG_NAME_CONST, K: "type", V: "element"},
-			{XMLName: XML_TAG_NAME_CONST, K: "subtype", V: "ms"},
-			{XMLName: XML_TAG_NAME_CONST, K: "id", V: signal.Name.Value},
-			{XMLName: XML_TAG_NAME_CONST, K: "direction", V: directionString},
-		},
-	}
 }
 
 // mapUnanchoredMainSignals processes all main signals for which no unique Node could be determined.
@@ -239,20 +214,8 @@ func searchUnanchoredMainSignal(
 	knoten Spurplanknoten,
 	isFalling bool,
 ) {
-	if len(*anchors) == 0 {
-		fmt.Print("Could not find anchors! \n")
-		return
-	}
-	if len(*anchors) == 1 {
-		fmt.Print("Could not find enough anchors! \n")
-		// TODO: Node not found, find closest mapped Node and work from there
-		return
-	}
-
-	directionString := "falling"
 	signals := knoten.HauptsigF
 	if !isFalling {
-		directionString = "rising"
 		signals = knoten.HauptsigS
 	}
 
@@ -265,129 +228,16 @@ func searchUnanchoredMainSignal(
 			continue
 		}
 
-		*nodeIdCounter++
-		newSignalNode := OSMUtil.Node{
-			Id:  strconv.Itoa(*nodeIdCounter),
-			Lat: maxNode.Lat,
-			Lon: maxNode.Lon,
-			Tag: []*OSMUtil.Tag{
-				{XMLName: XML_TAG_NAME_CONST, K: "type", V: "element"},
-				{XMLName: XML_TAG_NAME_CONST, K: "subtype", V: "ms"},
-				{XMLName: XML_TAG_NAME_CONST, K: "id", V: signal.Name.Value},
-				{XMLName: XML_TAG_NAME_CONST, K: "direction", V: directionString},
-			},
-		}
-		OSMUtil.InsertNewNodeWithReferenceNode(osmData, &newSignalNode, maxNode)
-	}
-}
-
-// findBestOSMNode determines a pair of anchors based on which a new Node is searched.
-// Based on those, a new Node is then determined.
-func findBestOSMNode(
-	osmData *OSMUtil.Osm,
-	anchors *map[float64]([]*OSMUtil.Node),
-	kilometrage float64,
-) (*OSMUtil.Node, error) {
-	sortedAnchors := sortAnchors(anchors, kilometrage)
-
-	nearest, secondNearest := sortedAnchors[0], sortedAnchors[1]
-
-	anchor1 := ((*anchors)[nearest])[0]
-	anchor2 := ((*anchors)[secondNearest])[0]
-	distance1 := math.Abs(nearest - kilometrage)
-	distance2 := math.Abs(secondNearest - kilometrage)
-
-	newNode, err := findNewNode(
-		osmData,
-		anchor1,
-		anchor2,
-		distance1,
-		distance2,
-	)
-
-	if err == nil {
-		return newNode, nil
-	}
-
-	newAnchorCounter := 2
-	for err != nil && newAnchorCounter < len(sortedAnchors) {
-		innerError := errors.Unwrap(err)
-		errorParts := strings.Split(innerError.Error(), ": ")
-		if errorParts[0] != "insufficient anchor" {
-			return nil, errors.Wrap(err, "failed to find OSM-node")
-		}
-
-		faultyNodeID := strings.ReplaceAll(errorParts[1], " ", "")
-
-		if faultyNodeID == anchor1.Id {
-			nearest = sortedAnchors[newAnchorCounter]
-			anchor1 = ((*anchors)[nearest])[0]
-			distance1 = math.Abs(nearest - kilometrage)
-			newAnchorCounter++
-		} else {
-			secondNearest = sortedAnchors[newAnchorCounter]
-			anchor2 = ((*anchors)[secondNearest])[0]
-			distance2 = math.Abs(secondNearest - kilometrage)
-			newAnchorCounter++
-		}
-		newNode, err = findNewNode(
+		newSignalNode := createNewHauptsignal(
+			nodeIdCounter,
+			maxNode,
+			signal,
+			isFalling,
+		)
+		OSMUtil.InsertNewNodeWithReferenceNode(
 			osmData,
-			anchor1,
-			anchor2,
-			distance1,
-			distance2,
+			&newSignalNode,
+			maxNode,
 		)
 	}
-
-	if newAnchorCounter == len(sortedAnchors) {
-		return nil, errors.New("failed to find suitable anchors")
-	}
-
-	return newNode, nil
-}
-
-// sortAnchors sorts the kilometrage-keys of the anchor-map.
-// The sort is based on least distance to the provided 'kilometrage'.
-func sortAnchors(
-	anchors *map[float64]([]*OSMUtil.Node),
-	kilometrage float64,
-) []float64 {
-	anchorKeys := []float64{}
-	for anchorKey := range *anchors {
-		anchorKeys = append(anchorKeys, anchorKey)
-	}
-
-	sort.SliceStable(anchorKeys, func(i, j int) bool {
-		floatKilometrage1, floatKilometrage2 := anchorKeys[i], anchorKeys[j]
-		return math.Abs(kilometrage-floatKilometrage1) < math.Abs(kilometrage-floatKilometrage2)
-	})
-
-	return anchorKeys
-}
-
-// formatKilometrageStringInFloat formats the kilometrage and parses them to a float64.
-// Kilometrages in the DB-data always use a comma for decimal separation and sometimes, a plus sign is present.
-// This plus sign is parsed in a way, that e.g. '5,000+0,150' becomes '5.15'.
-func formatKilometrageStringInFloat(
-	in string,
-) (float64, error) {
-	split := strings.Split(in, "+")
-	if len(split) == 1 {
-		return strconv.ParseFloat(
-			strings.ReplaceAll(in, ",", "."),
-			64)
-	}
-
-	out := 0.0
-	for _, splitPart := range split {
-		floatPart, err := strconv.ParseFloat(
-			strings.ReplaceAll(splitPart, ",", "."),
-			64)
-		if err != nil {
-			return 0, errors.Wrap(err, "failed to parse kilometrage: "+in)
-		}
-		out += floatPart
-	}
-
-	return out, nil
 }
