+++ generate-osm/osm-utils/osm-xml_structure.go
